- correct tests - update productOfChars primitive from int to long - add check for word length of 1 before creating lower bound and upper bound sub sets - abstract hardcoded numbers for acceptable edit distance and upper and lower bound subsets into variables - refactor to remove the utilities method 'removeDuplicates' - rename variables in the levenshteinDistance calculating method - clean imports and spacing

+++ src/main/java/entity/UniverseEntity.java
@@ -4,7 +4,7 @@ import utils.Utilities;
 
 public class UniverseEntity {
 
-	final int productOfChars;
+	final long productOfChars;
 	final String originalEntry;
 	
 	public UniverseEntity(String originalEntry) {
@@ -12,11 +12,11 @@ public class UniverseEntity {
 		productOfChars = calculateProductOfChars(originalEntry);
 	}
 
-	protected int calculateProductOfChars(String originalEntry) {
+	protected long calculateProductOfChars(String originalEntry) {
 	  return Utilities.calculateProductOfChars(originalEntry);
 	}
 	
-	public int getProductOfChars()
+	public long getProductOfChars()
 	{
 		return productOfChars;
 	}
